* Removing implicit cast to prevent cast exceptions. * Fix FindLibrary to only return a match when the library is non-null * Print the package path correctly.
