+++ src/NuGet.ProjectModel/LockFileTargetLibrary.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using NuGet.Packaging.Core;
+using NuGet.Versioning;
+
+namespace NuGet.ProjectModel
+{
+    public class LockFileTargetLibrary
+    {
+        public string Name { get; set; }
+
+        public NuGetVersion Version { get; set; }
+
+        public IList<PackageDependency> Dependencies { get; set; } = new List<PackageDependency>();
+
+        public IList<string> FrameworkAssemblies { get; set; } = new List<string>();
+
+        public IList<LockFileItem> RuntimeAssemblies { get; set; } = new List<LockFileItem>();
+
+        public IList<LockFileItem> ResourceAssemblies { get; set; } = new List<LockFileItem>();
+
+        public IList<LockFileItem> CompileTimeAssemblies { get; set; } = new List<LockFileItem>();
+
+        public IList<LockFileItem> NativeLibraries { get; set; } = new List<LockFileItem>();
+    }
+}
